+++ core/terraform/network/variables.tf
@@ -13,8 +13,6 @@ variable "core_address_space" {
 variable "arm_environment" {
   type = string
 }
-variable "route_table_id" {
-  type        = string
-  description = "The ID of the route table to associate with subnets"
-  default     = null
+variable "firewall_force_tunnel_ip" {
+  type = string
 }