+++ core/terraform/routetable.tf
@@ -1,22 +1,27 @@
-resource "azurerm_route_table" "fw_tunnel_rt" {
-  count                         = var.firewall_force_tunnel_ip != "" ? 1 : 0
-  name                          = "rt-fw-tunnel-${var.tre_id}"
-  resource_group_name           = azurerm_resource_group.core.name
-  location                      = azurerm_resource_group.core.location
-  bgp_route_propagation_enabled = true
-  tags                          = local.tre_core_tags
-  lifecycle { ignore_changes = [tags] }
-
-  route {
-    name                   = "ForceTunnelRoute"
-    address_prefix         = "0.0.0.0/0"
-    next_hop_type          = "VirtualAppliance"
-    next_hop_in_ip_address = var.firewall_force_tunnel_ip
-  }
+moved {
+  from = azurerm_route_table.rt
+  to   = module.network.azurerm_route_table.rt
 }
 
-resource "azurerm_subnet_route_table_association" "rt_fw_tunnel_subnet_association" {
-  count          = var.firewall_force_tunnel_ip != "" ? 1 : 0
-  subnet_id      = module.network.azure_firewall_subnet_id
-  route_table_id = azurerm_route_table.fw_tunnel_rt[0].id
+moved {
+  from = azurerm_route_table.fw_tunnel_rt
+  to   = module.network.azurerm_route_table.fw_tunnel_rt
+}
+
+import {
+  to = azurerm_route.default_route
+  id = "/subscriptions/${data.azurerm_client_config.current.subscription_id}/resourceGroups/${azurerm_resource_group.core.name}/providers/Microsoft.Network/routeTables/${module.network.route_table_name}/routes/DefaultRoute"
+}
+
+resource "azurerm_route" "default_route" {
+  name                   = "DefaultRoute"
+  resource_group_name    = azurerm_resource_group.core.name
+  route_table_name       = module.network.route_table_name
+  address_prefix         = "0.0.0.0/0"
+  next_hop_type          = "VirtualAppliance"
+  next_hop_in_ip_address = module.firewall.private_ip_address
+
+  depends_on = [
+    module.firewall
+  ]
 }